refactor: split InspectorPanels.tsx into focused modules — reduce 538-line file to focused per-panel components
